Corrige une erreur dans la suppression des éléments ajoutables
