+++ src/vues/fragments/elementsAjoutables.pug
@@ -14,3 +14,5 @@ mixin elementsAjoutables({ identifiantConteneur, nom, valeurExemple = '', donnee
     class = 'nouvel-item',
     id = 'ajout-element-' + nom
   )= texteLienAjouter
+
+  script(type = 'module', src = '/statique/modules/elementsAjoutables.js')
